util/mock: move mocks to mock package
